+++ impexp-core/src/main/java/org/citydb/ade/kmlExporter/ADEKmlExportExtension.java
@@ -25,10 +25,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.citydb.modules.kml.ade;
+package org.citydb.ade.kmlExporter;
 
-public interface ADEKmlExporter {
-	String getPointAndCurveQuery(int lod);
-	String getSurfaceGeometryQuery(int lod);
-	String getSurfaceGeometryRefIdsQuery(int lod);
+public interface ADEKmlExportExtension extends ADEBalloonExtension {
+	ADEKmlExportManager createADEKmlExportManager();
 }
