+++ impexp-core/src/main/java/org/citydb/ade/kmlExporter/ADEKmlExportHelper.java
@@ -25,12 +25,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.citydb.modules.kml.ade;
+package org.citydb.ade.kmlExporter;
 
 import org.citydb.database.adapter.AbstractDatabaseAdapter;
-import org.citydb.modules.kml.database.Queries;
+
+import java.util.List;
 
 public interface ADEKmlExportHelper {
 	AbstractDatabaseAdapter getDatabaseAdapter();
 	Queries getSQLQueries();
+	List<ADEKmlExportManager> getADEKmlExportManagers();
 }
