+++ impexp-kml-collada-plugin/src/main/java/org/citydb/modules/kml/database/KmlExporterManager.java
@@ -39,6 +39,7 @@ import net.opengis.kml._2.RegionType;
 import net.opengis.kml._2.ViewRefreshModeEnumType;
 import org.citydb.ade.ADEExtension;
 import org.citydb.ade.ADEExtensionManager;
+import org.citydb.ade.kmlExporter.*;
 import org.citydb.concurrent.WorkerPool;
 import org.citydb.config.Config;
 import org.citydb.config.project.kmlExporter.DisplayForm;
@@ -49,8 +50,6 @@ import org.citydb.event.EventDispatcher;
 import org.citydb.event.global.CounterEvent;
 import org.citydb.event.global.CounterType;
 import org.citydb.log.Logger;
-import org.citydb.modules.kml.ade.*;
-import org.citydb.modules.kml.controller.KmlExportException;
 import org.citydb.modules.kml.util.BalloonTemplateHandler;
 import org.citydb.modules.kml.util.CityObject4JSON;
 import org.citydb.modules.kml.util.ExportTracker;
